deploy/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-multipart==0.0.6

# AI Model Dependencies
tensorflow==2.20.0
numpy==2.3.3
Pillow==11.3.0
